Move AuthenticatedRoute into container
